+++ .gitignore
@@ -25,3 +25,4 @@ front/dist/
 # Only for back-end
 # #############################################################################
 
+back/build
