+++ back/package.json
@@ -11,11 +11,12 @@
     "express-graphql": "^0.12.0",
     "graphql": "^15.8.0",
     "jsonwebtoken": "^8.5.1",
-    "mysql2": "^2.3.3"
+    "mysql2": "^2.3.3",
+    "typescript": "^4.6.3"
   },
   "scripts": {
-    "dev": "NODE_ENV=development node ./src/main.js",
-    "start": "NODE_ENV=production node ./src/main.js",
+    "dev": "NODE_ENV=development node ./build/main.js",
+    "start": "NODE_ENV=production node ./build/main.js",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest"
   },
   "type": "module"
