+++ back/tsconfig.json
@@ -1,0 +1,20 @@
+{
+    "compilerOptions": {
+        "module": "es6",
+        "target": "es6",
+        "noImplicitAny": false,
+        "sourceMap": false,
+        "rootDir":"./src",
+        "outDir":"./build",
+        "watch":false,
+        "allowJs": true,
+        "moduleResolution": "node"
+    },
+    "include":[
+        "./src/**/*"
+    ],
+    "exclude":[
+        "views",
+        "static"
+    ]
+}
